refactor: group all protocols in single file
