+++ src/data/use-cases/add-account/db-add-account.protocols.ts
@@ -1,0 +1,3 @@
+export * from '../../../domain/use-cases/add-account';
+export * from '../../../domain/models/account';
+export * from '../../protocols/encrypter';
