+++ src/data/use-cases/add-account/db-add-account.ts
@@ -1,6 +1,4 @@
-import { AddAccount, AddAccountModel } from '../../../domain/use-cases/add-account';
-import { AccountModel } from '../../../domain/models/account';
-import { Encrypter } from '../../protocols/encrypter';
+import { AddAccount, AddAccountModel, AccountModel, Encrypter } from './db-add-account.protocols';
 
 export class DbAddAccount implements AddAccount {
   private readonly encrypter: Encrypter;
